langserve_backend/requirements.txt
fastapi
uvicorn
langserve
langgraph
langchain
requests
streamlit
python-dotenv
sse_starlette>=1.6.5

